Modify SRAAACA in splinters so the partner is captain
Rather than try to communicate wasted values in the suit partner is
short in, just make the partner of the splintering player captain. Since
overall values are less likely to matter with very distributional hands,
this also allows the captain to bypass the strength response and
directly ask about suit holdings.