Adjust balanced positives and add ask for 4 card majors
4-4 major fits will often be better than 3NT, and we are mostly unable
to find them when opener also has no 5 card suit (since the first suit
bid would be expected to be 5). Narrowing the 2NT range helps with the
fact that there is significantly less bidding space there, while the 1NT
responses tended to have extra that could be useful for slam bidding.

In the future, we might consider elaborating the system here. For
example, symmetric relay has the idea of "zooming" which is responding
to the next question before it's asked when you're giving the top answer
to the current question. We could do something similar here with after
1C-1NT-2C, 2NT and higher all show both majors, with each bid showing a
point range.