Adjust 2C responses
I tried running simulations with 2D as a potentially garbage stayman
(and always responding at the lowest level when having a major), but the
frequency of hitting a 4-4 major fit was lower than the frequency of
going down an extra trick in clubs, and responder with 4-4 majors will
often do well to defend.

However, the 2NT invite showing 5S4H looked to be a winner when it came
up, so adding that and moving the natural 2NT invite into 2D a la
non-promissory Stayman.

Additionally, this change defines the club raises as preemptive, and the
3 level bids as natural game forces.