Make new suits after weak 2s nonforcing (#69)
Because our style opens weak 2s very light, hands with interest in game
are quite rare. Additionally, since we do not have strict requirements
on suit quality, hands where responder wants to be able to run to a
different suit even in the absence of a double are quite common.